combine setup and run into 1 file
