+++ libs/auto-docs/test/setup.test.ts
@@ -4,6 +4,10 @@ import { Api } from '@modtree/repo-api'
 import '@modtree/test-env/jest'
 import { User, Degree, Graph } from '@modtree/entity'
 import { init } from '../src/init'
+import { routes } from '../src/routes'
+import { postman } from '../src/postman'
+import fs from 'fs'
+import { join } from 'path'
 
 const dbName = 'modtree'
 const db = getSource(dbName)
@@ -16,6 +20,8 @@ let d1: Degree
 let d2: Degree
 let g1: Graph
 let g2: Graph
+
+jest.setTimeout(60000)
 
 beforeAll(() =>
   setup(db).then(() => {
@@ -97,5 +103,47 @@ describe('properties checks', () => {
         if (Array.isArray(value)) expect(value.length).toBeGreaterThan(0)
       })
     })
+  })
+})
+
+describe('samples', () => {
+  const all: any[] = []
+
+  test.each(routes)('%p %p', async (method, route, routeInfo) => {
+    // unpack data
+    const url = routeInfo.url
+    const params = routeInfo.params
+
+    // make request
+    const res = await postman.send(method, url, { params })
+
+    // truncate response data if necessary
+    let response
+    if (Array.isArray(res.data) && res.data.length > 3) {
+      // if the response is an array, only give at most 3 entries
+      response = [res.data[0], res.data[1], res.data[2]]
+    } else {
+      // otherwise just leave it
+      response = res.data
+    }
+
+    // transform data
+    const data = {
+      method: res.config.method,
+      url: res.config.url,
+      res: response,
+    }
+
+    all.push(data)
+
+    expect(1).toEqual(1)
+  })
+
+  test('write to file', () => {
+    const rootDir = join(__dirname, '../../..')
+    const json = JSON.stringify(all, null, 2)
+    fs.writeFileSync(join(rootDir, 'references/api.json'), json)
+
+    expect(1).toEqual(1)
   })
 })
