+++ package.json
@@ -25,8 +25,7 @@
     "i": "bash ./scripts/yarn.sh inverse",
     "heroku-postbuild": "nx build server",
     "version": "auto-changelog -p && git add --all",
-    "auto:setup": "nx test auto-docs --testPathPattern setup",
-    "auto:run": "nx test auto-docs --testPathPattern run"
+    "auto:setup": "nx test auto-docs --testPathPattern setup"
   },
   "dependencies": {
     "@auth0/nextjs-auth0": "^1.9.0",
